Mass-merged the IP address/port code cleanup
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@10309 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/login_sql/login.h
@@ -11,13 +11,9 @@
 #define LAN_CONF_NAME "conf/subnet_athena.conf"
 
 #ifndef SQL_DEBUG
-
-#define mysql_query(_x, _y) mysql_query(_x, _y)
-
+	#define mysql_query(_x, _y) mysql_query(_x, _y)
 #else 
-
-#define mysql_query(_x, _y)  debug_mysql_query(__FILE__, __LINE__, _x, _y)
-
+	#define mysql_query(_x, _y)  debug_mysql_query(__FILE__, __LINE__, _x, _y)
 #endif
 
 #define PASSWORDENC		3	// A definition is given when making an encryption password correspond.
@@ -42,8 +38,8 @@ struct mmo_account {
 
 struct mmo_char_server {
 	char name[20];
-	long ip;
-	short port;
+	uint32 ip;
+	uint16 port;
 	int users;
 	int maintenance;
 	int new_;
